Move variant generation code to separate files
